Rename index.ts to config.ts (and introduce a new smaller index file)
